Make return route more flexible and use for lock investmnet
